% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageCut.R
\name{ageCut}
\alias{ageCut}
\title{Cut all continuous age values to categorical.}
\usage{
ageCut(x, interval = 5)
}
\arguments{
\item{x}{a vector which have numeric age values.}

\item{interval}{interval of age, between 5 or 10.}
}
\description{
ageCut cut all numeric age values to categorical.
 You can choose interval 5 or 10.
}
\examples{
 cut1 <- ageCut(c(12,21,31,11,22,53,62,34,55))
 cut2 <- ageCut(c(12,21,31,11,22,53,62,34,55), 10)

}
