% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexBlender.R
\name{hexBlender}
\alias{hexBlender}
\title{Blend two colors using hex codes.}
\usage{
hexBlender(color1, color2, alpha = 0.5)
}
\arguments{
\item{color1}{character value of hex code. no default.}

\item{color2}{character value of hex code. no default.}

\item{alpha}{float value from over 0 and under 1. Default value set as 0.5.}
}
\description{
hexBlender function blend two colors of hex code and generate new hex code.
}
\examples{
 new_hex <- hexBlender('#FF0000', '#000000')

}
