% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordComb.R
\name{keywordComb}
\alias{keywordComb}
\title{Count all combinations of elements in vector and make it as dataframe.}
\usage{
keywordComb(x, m = 2, remove_dup = T)
}
\arguments{
\item{x}{a vector which have more than m element(length(x) >= m).}

\item{m}{a number of combination. Default set as 2.}

\item{remove_dup}{if TRUE, use only unique elements in vectors(which means remove duplicates).}
}
\description{
keywordComb count all combinations of elements in vector.
 Usually, this function used for network analysis when calculate weight between nodes(or vertexs).
 Please refer more details from github or example.
}
\examples{
 comb2 <- keywordComb(as.character(c(1,1,2,2,1,2,1,2,1,5,5,5,3,6,5,3,5)), 2)
 comb3 <- keywordComb(as.character(c(1,1,2,2,1,2,1,2,1,5,5,5,3,6,5,3,5)), 3, remove_dup = T)

}
