% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordVec.R
\name{keywordVec}
\alias{keywordVec}
\title{Count elements in vector and make it as one string.}
\usage{
keywordVec(x, reverse = F, limit = NULL)
}
\arguments{
\item{x}{vector of characters.}

\item{reverse}{logical. If true, convert string value to vectors.}

\item{limit}{.}
}
\description{
keywordVec count all elements in vector. Please refer example.
}
\examples{
 count <- keywordVec(as.character(c(1,1,2,2,1,2,1,2,1,5,5,5,3,6,5,3,5)))
 reverse <- keywordVec(count, reverse = TRUE)

}
