% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today6d.R
\name{today6d}
\alias{today6d}
\title{Generate 6-digits string of today.}
\usage{
today6d(with_time = FALSE)
}
\description{
today6d generate 6-digits string which indicate current day.
}
\examples{
 d1 <- today6d()
 d2 <- today6d(with_time = T)

}
