% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_dataframe.R
\name{xml_to_dataframe}
\alias{xml_to_dataframe}
\title{Convert object which class is 'xml_nodeset' to data.frame.}
\usage{
xml_to_dataframe(nodeset)
}
\arguments{
\item{nodeset}{a dataframe.}
}
\description{
xml_to_dataframe() convert xml_nodeset object to data.frame.
}
\examples{
xml <- "<?xml  version='1.0' encoding='UTF-8'?>
<response>
  <header>
    <resultMsg>NORMAL SERVICE</resultMsg>
  </header>
  <body>
    <items>
      <item>
        <num>1</num>
        <date>20160413</date>
        <title>test_title1</title>
        <code>0</code>
        <time>08:23</time>
      </item>
      <item>
        <num>2</num>
        <date>20160414</date>
        <title>test_title2</title>
        <code>2</code>
        <time>22:23</time>
      </item>
      <item>
        <num>3</num>
        <date>20200415</date>
        <title>test_title3</title>
        <code>7</code>
        <time>20:47</time>
      </item>
    </items>
    <numOfRows>3</numOfRows>
    <pageNo>1</pageNo>
    <totalCount>3</totalCount>
  </body>
</response>"

doc <- xml2::read_xml(xml)
items <- xml2::xml_find_all(doc, '//item')
df <- xml_to_dataframe(items)

}
